autoGenerations/requirements.txt
sqlalchemy
requests
pillow